requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.15.1,<0.16
django-taggit==4.0.0
django-cors-headers==3.13.0
psycopg2>=2.8.6,<2.9
pillow>=10.0.0,<11.0.0
stripe==13.0.1